added lowest-level sorting of particles, minor fixes to the ic-reader, and extension of the copyright statement.
Former-commit-id: 8f1de71456a8c10e009da2d48cbc7683106418c4

+++ src/part.h
@@ -31,13 +31,13 @@ struct part {
     float h;
     
     /* Particle time-step. */
-    int dt;
+    float dt;
     
     /* Particle mass. */
     float mass;
     
     /* Particle ID. */
-    unsigned long id;
+    unsigned long long id;
     
     /* Particle position. */
     double x[3];
